Fix Particle Overlap
In order to avoid ever more overlapping particles, check in findCollisionTime whether the particles overlap and if yes perform a collision with collisionTime = 0, that is change the velocities only. The time is only driven by valid collsions. In order to avoid infinite loops, the flag in main lastIsOverlapping ensures, there is always at least one valid collision, that is there is a positional change.